packages/frontend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fastset-frontend"
version = "0.1.0"
description = "FastSet BI Platform Frontend"
requires-python = ">=3.8"
dependencies = [
    "python-fasthtml>=0.12.25",
]

[tool.uv]
dev-dependencies = [
]

[tool.hatch.build.targets.wheel]
packages = ["src/frontend"]
